+++ src/documentation/pages/components-wip/forms/radio-button/examples/_optional.vue
@@ -5,12 +5,13 @@
       | Use <code>optional</code> to help emphasize a group of radio button options are not required and can be skipped.
   template(#example)
     fieldset
-      legend(class="chi-label") Select an option
-        abbr(class="chi-label__optional" aria-label="Optional field") (optional)
-      div.chi-form__item
-        chi-radio-button(id="example__optional_opt_1" label="Option 1" name="radios")
-      div.chi-form__item
-        chi-radio-button(id="example__optional_opt_2" label="Option 2" name="radios")
+      legend.chi-label
+        | Select an option
+        abbr.chi-label__optional(aria-label="Optional field") (optional)
+      .chi-form__item.-mb--1  
+        chi-radio-button(id="example__optional_opt_1" label="Option 1" name="example__optional-radios")
+      .chi-form__item
+        chi-radio-button(id="example__optional_opt_2" label="Option 2" name="example__optional-radios")
   template(#code-webcomponent)
     Copy(lang="html" :code="codeSnippets.webcomponent" class="html")
   template(#code-htmlblueprint)
@@ -42,7 +43,7 @@ import { Vue } from 'vue-facing-decorator';
     Select an option
     <abbr class="chi-label__optional" aria-label="Optional field">(optional)</abbr>
   </legend>
-  <div class="chi-form__item">
+  <div class="chi-form__item -mb--1">
     <chi-radio-button id="example__optional_opt_1" label="Option 1" name="radios"></chi-radio-button>
   </div>
   <div class="chi-form__item">
@@ -54,7 +55,7 @@ import { Vue } from 'vue-facing-decorator';
     Select an option
     <abbr class="chi-label__optional" aria-label="Optional field">(optional)</abbr>
   </legend>
-  <div class="chi-form__item">
+  <div class="chi-form__item -mb--1">
     <div class="chi-radio">
       <input class="chi-radio__input" type="radio" name="radios" id="example__optional_opt_1">
       <label class="chi-radio__label" for="example__optional_opt_1">Option 1</label>